+++ src/Ycs/lib0/Decoding/IntDiffDecoder.cs
@@ -9,18 +9,18 @@ using System.IO;
 namespace Ycs
 {
     /// <seealso cref="IntDiffEncoder"/>
-    internal class IntDiffDecoder : AbstractStreamDecoder<int>
+    internal class IntDiffDecoder : AbstractStreamDecoder<long>
     {
-        private int _state;
+        private long _state;
 
-        public IntDiffDecoder(Stream input, int start, bool leaveOpen = false)
+        public IntDiffDecoder(Stream input, long start, bool leaveOpen = false)
             : base(input, leaveOpen)
         {
             _state = start;
         }
 
         /// <inheritdoc/>
-        public override int Read()
+        public override long Read()
         {
             CheckDisposed();
